Check for newer file when concatenating tests

+++ scripts/concat_tests.sh
@@ -2,15 +2,25 @@
 
 repo="$(dirname $(dirname $(readlink -m $0)))"
 cd ${repo}
-output_file=concat_tests.gdb
-rm -f "${output_file}"
-find ${repo}/tests -name \*.gdb | while read filename; do
-	{
-		echo "# ${filename}" | sed "s|${repo}/||g"
-		cat "${filename}"
-		echo ""
-	} >> "${output_file}"
-done
+output_file=${repo}/tests/concat_tests.gdb
+if [[ ! -f ${output_file} ]]; then
+	cond=1
+else
+	nb_new_file=$(find ${repo}/tests -name \*.gdb -newer ${output_file} | wc -l)
+	echo "# ${nb_new_file} newer files" > /dev/stderr
+	cond=${nb_new_file}
+fi
+if [[ $cond -ne 0 ]]; then
+	echo "# creating ${output_file} " > /dev/stderr
+	rm -f "${output_file}"
+	find ${repo}/tests -name \*.gdb | while read filename; do
+		{
+			echo "# ${filename}" | sed "s|${repo}/||g"
+			cat "${filename}"
+			echo ""
+		} >> "${output_file}"
+	done
+fi
 if [[ -f "${output_file}" ]]; then
 	readlink -m "${output_file}"
 else
